doc: Document the primary types used by the scanner
