+++ src/scanner/slowFunctionCall.ts
@@ -1,5 +1,5 @@
 import { Event } from '@appland/models';
-import { ScopeName } from 'src/types';
+import { AssertionSpec, ScopeName } from 'src/types';
 import Assertion from '../assertion';
 import { toRegExp } from './util';
 
@@ -48,4 +48,9 @@ function scanner(options: Options): Assertion {
   );
 }
 
-export default { scope: 'root' as ScopeName, enumerateScope: true, Options, scanner };
+export default {
+  scope: 'root' as ScopeName,
+  enumerateScope: true,
+  Options,
+  scanner,
+} as AssertionSpec;
